implemented the mobile menu
this was interesting. designed it to leave the main nav bar in
place and overaly another element on top and anchored in the top
right corner. the log and site name are styled the same as the
main nav and the hamburger button is changed to x button in same
location. i am pleased with this for now, let's see how it works
when i deploy to prod and test on mobile device.