(#36) add sidebar to reference content layout
The goal of this change was to add a layout having a left sidebar for links to
articles within a section. As part of this goal some additional changes were
made to improve the site in pursuit of this goal. The following changes have
been applied:

* Added filter to debug data during development
* Removed the terminal collection override and replaced with a generic sections
  collection that is data driven.
* Renamed base layout to full_width to make its intent more obvious.
* Added left_sidebar layout which includes its own mobile menu separate from
  the page level mobile menu.
* Added data driven macro to generate list of links for a given collection.
* Pulled out footer from base layout into partial for reuse in other layouts.
* Changed the state variable name of the page header mobile menu.
* Removed hard coded links from the termianl index page.
* Added concetp of section for use in reference collections.
* Removed most nunjucks block sections only keeping the ones actually used.

Overall got a better understanding of 11ty collections and how to create a more
data driven layout. Learned more about Tailwind use for layouts, styling, and
transitions. Had to learn about conflicts using Tailwind transtions with Alpine. When
Alpine is controlling the visibility of an element then you need to use the
Alpine x-transition attributes to control the transition. You can use Tailwind
classes as values in those attributes. You cannot however use those
transitions on the element itself, won't work.