(#37) implement blog post list page
Big change here is renaming Blog to Bytes and post(s) to byte(s). For better
or worse I did not want to use the typical names of Blog or post. In addition,
I miss the old Twitter and still want to post short form content previously
known as Tweets only I am not going to name them Tweets. So I came up with the
idea of Bits and Bytes. Bits will be the same as Tweets and a separate part of
the site. Writing longer form content, or traditional blog, will be Bytes.

Created a new layout for Bytes extending Full Width layout. I added a block for
main so that the Byte layout could include the title and publication date of
the Byte. The list of Bytes is in descending order by publication date grouped
by year. Each year in which a Byte was published will be used as a heading for
those Bytes. No pagination, this list will be all available Bytes. This should
be fine for years to come.

In addition to the basic feature of rendering a list of blog posts, added date
filters to parse and format dates replacing previous shortcode and added posts
collection to provide posts in descending order by publication date.